outlined a task plan in app component and refactoring code
