+++ src/components/Voting/Dislikes/Dislikes.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import s     from './Dislikes.module.scss'
+import s     from '../Voting.module.scss'
 
 import { TDataObj }  from '../../../redux/voting/types'
 import { TDislikes } from './types'
@@ -12,10 +12,10 @@ const Dislikes: React.FC<TDislikes> = ({ unlikeData, status }) => {
 	return (
 		<>
 			{ noItemsBoolean && <div className='noItemFound '><span>No item found</span></div> }
-			<div className={ s.dislikes }>
+			<div className={ s.voting__items }>
 				{ unlikeData?.map((el: TDataObj) => {
 					return (
-						<div className={ s.dislikes__img_wr } key={ el?.id }>
+						<div className={ s.voting__itemsImg_wr } key={ el?.id }>
 							<img src={ el?.url } alt='image'/>
 						</div>
 					)
