+++ src/components/Voting/Voting.module.scss
@@ -45,4 +45,35 @@
       border-radius: 20px;
     }
   }
+
+  &__items {
+    display: grid;
+    grid-auto-rows: 140px;
+    grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+    grid-gap: 20px;
+    margin-bottom: 20px;
+    min-height: 360px;
+
+    &Img_wr {
+
+      img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+        border-radius: 20px;
+      }
+
+      &:first-child,
+      &:nth-child(10n+1),
+      &:nth-child(10n-2) {
+        grid-row: span 2;
+      }
+
+      &:nth-child(5n-1),
+      &:nth-child(10n-1) {
+        grid-column: span 2;
+        grid-row: span 2;
+      }
+    }
+  }
 }
